requirements.txt
sympy==1.13.2
manim==0.18.1
rich==13.7.1
typer==0.12.5
pydantic==2.9.2
python-dotenv==1.0.1
fastapi
uvicorn
